pyproject.toml
[tool.poetry]
name = "skillplot"
version = "0.1.0"
description = ""
authors = ["Felix Warmuth <felixwarmuth@mailbox.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
click = "^8.1.7"
loguru = "^0.7.2"
matplotlib = "^3.8.0"
pandas = "^2.1.1"
pyyaml = "^6.0.1"
ruamel-yaml = "^0.17.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
skillplot = "skillplot.__main__:main"